src/requirements.txt
ffmpeg-python
einops
yt-dlp
zipp
psycopg2-binary
pgvector
torchvision
aiohttp==3.9.1
python-dateutil==2.8.2
pysrt==1.1.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
python-slugify==8.0.1
python-utils==3.8.1
Flask==2.1.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-Pydantic==0.10.0
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.38.3
Pillow==9.1.1
psycopg2-binary==2.9.3
gunicorn==20.1.0
opencv-python
pyping==0.0.6
python-keycloak==2.6.0
requests>=2.23.0
redis==4.3.5
pydantic==1.10.2
gevent==22.10.2
XlsxWriter==3.1.2
Werkzeug==2.2.2
oauthlib==3.0.1
av
transformers
torch
easydict
scikit-learn